Add remaining skill tables
I'm really not sure about these... modeling the restrictions could be
done in a handful of ways, but I don't know what the most _usable_ of
those ways would be. I've tentatively listed out all the move types,
weapon types, and color types as bitfields (with the notable exception
of "colored bow", since apparently that's a relevant thing).

I could opt to make a Restrictions table and that lets me use a single
fkey for this, but that seems really useless for all but the "query by
hand" use-case.

I could opt to make separate mapping tables to model the restrictions.
This is a bit problematic, since we're spreading out these related items
(move type, weapon type, skill type) across multiple tables, and I'm not
keen on having nullable, mutually-exclusive fkeys for a single row.

I could enumerate the color-weapon pairs and drop the color-specific
bits. This may be the most appealing solution to me, I'm afraid.

I'll think on it some more.